:recycle: Chore: Refactor out actions todo
