+++ examples/esbuild/package.json
@@ -2,9 +2,7 @@
 	"name": "@tevm/example-esbuild",
 	"version": "0.0.1",
 	"private": true,
-	"contributors": [
-		"Will Cory <willcory10@gmail.com>"
-	],
+	"contributors": ["Will Cory <willcory10@gmail.com>"],
 	"type": "module",
 	"scripts": {
 		"//build": "bun run build:app",
