+++ packages/blockchain/src/BaseChain.ts
@@ -7,9 +7,9 @@ import type { ChainOptions } from './ChainOptions.js'
  * Internal implementation of blockchain used to implement the public interface
  */
 export type BaseChain = {
-  /**
-   * Options passed into factory function
-   */
+	/**
+	 * Options passed into factory function
+	 */
 	options: ChainOptions
 	common: Common
 	/**
