+++ packages/blockchain/src/ChainOptions.ts
@@ -6,27 +6,27 @@ import type { BlockTag } from 'viem'
  * Options passed into `createChain` to initialize a Chain object
  */
 export type ChainOptions = {
-  /**
-   * A Common instance
-   */
+	/**
+	 * A Common instance
+	 */
 	common: Common
-  /**
-   * Override the genesis block. If fork is provided it will be fetched from fork. Otherwise a default genesis is provided.
-   */
+	/**
+	 * Override the genesis block. If fork is provided it will be fetched from fork. Otherwise a default genesis is provided.
+	 */
 	genesisBlock?: Block
 	genesisStateRoot?: Uint8Array
-  /**
-   * Optional fork config for forking a live chain
-   */
+	/**
+	 * Optional fork config for forking a live chain
+	 */
 	fork?: {
-    /** 
-     * JSON-RPC url to fork
-     */
+		/**
+		 * JSON-RPC url to fork
+		 */
 		url: string
-    /**
-     * Optional block tag to fork
-     * Defaults to 'latest'
-     */
+		/**
+		 * Optional block tag to fork
+		 * Defaults to 'latest'
+		 */
 		blockTag?: BlockTag | bigint | `0x${string}`
 	}
 }
