+++ packages/blockchain/src/actions/delBlock.js
@@ -1,0 +1,8 @@
+/**
+ * @param {import('../BaseChain.js').BaseChain} baseChain
+ * @returns {import('@ethereumjs/blockchain').BlockchainInterface['delBlock']}
+ */
+export const delBlock = (baseChain) => (blockHash) => {
+	baseChain.blocks.delete(blockHash)
+	return Promise.resolve()
+}
