+++ packages/blockchain/src/actions/getIteratorHead.js
@@ -1,0 +1,15 @@
+/**
+ * @param {import('../BaseChain.js').BaseChain} baseChain
+ * @returns {import('@ethereumjs/blockchain').BlockchainInterface['getIteratorHead']}
+ */
+export const getIteratorHead =
+	(baseChain) =>
+	(name = 'vm') => {
+		const head = baseChain.blocksByTag.get(/** @type {import('viem').BlockTag}*/ (name))
+		if (!head) {
+			throw new Error(
+				`No block with tag ${name} exists. Current tags include ${[...baseChain.blocksByTag.keys()].join(',')}`,
+			)
+		}
+		return Promise.resolve(head)
+	}
