+++ packages/blockchain/src/createBaseChain.js
@@ -20,52 +20,52 @@ const KECCAK256_RLP = hexToBytes(KECCAK256_RLP_S)
  * @returns {Block}
  */
 const createGenesisBlock = (stateRoot, common) => {
-  const newCommon = common.copy()
-  newCommon.setHardforkBy({
-    blockNumber: 0,
-    td: BigInt(newCommon.genesis().difficulty),
-    timestamp: newCommon.genesis().timestamp ?? 0,
-  })
+	const newCommon = common.copy()
+	newCommon.setHardforkBy({
+		blockNumber: 0,
+		td: BigInt(newCommon.genesis().difficulty),
+		timestamp: newCommon.genesis().timestamp ?? 0,
+	})
 
-  /**
-   * @type {import('@tevm/block').HeaderData}
-   */
-  const header = {
-    ...newCommon.genesis(),
-    number: 0,
-    stateRoot,
-    ...(newCommon.isActivatedEIP(4895) ? { withdrawalsRoot: KECCAK256_RLP } : {}),
-  }
-  return Block.fromBlockData({ header, ...(newCommon.isActivatedEIP(4895) ? { withdrawals: [] } : {}) }, { common })
+	/**
+	 * @type {import('@tevm/block').HeaderData}
+	 */
+	const header = {
+		...newCommon.genesis(),
+		number: 0,
+		stateRoot,
+		...(newCommon.isActivatedEIP(4895) ? { withdrawalsRoot: KECCAK256_RLP } : {}),
+	}
+	return Block.fromBlockData({ header, ...(newCommon.isActivatedEIP(4895) ? { withdrawals: [] } : {}) }, { common })
 }
 /**
  * @param {import('./ChainOptions.js').ChainOptions} options
  * @returns {import('./BaseChain.js').BaseChain} Base chain object
  */
 export const createBaseChain = (options) => {
-  /**
-   * @type {import('./BaseChain.js').BaseChain}
-   */
-  const chain = {
-    options,
-    common: options.common,
-    blocks: new Map(),
-    blocksByTag: new Map(),
-    blocksByNumber: new Map(),
-    ready: () => genesisBlockPromise.then(() => true),
-  }
+	/**
+	 * @type {import('./BaseChain.js').BaseChain}
+	 */
+	const chain = {
+		options,
+		common: options.common,
+		blocks: new Map(),
+		blocksByTag: new Map(),
+		blocksByNumber: new Map(),
+		ready: () => genesisBlockPromise.then(() => true),
+	}
 
-  // Add genesis block and forked block to chain
-  const genesisBlockPromise = (async () => {
-    if (options.fork?.url) {
-      const block = await getBlockFromRpc(options.fork, options.common)
-      await putBlock(chain)(block)
-    } else {
-      await putBlock(chain)(
-        options.genesisBlock ?? createGenesisBlock(options.genesisStateRoot ?? EMPTY_STATE_ROOT, options.common),
-      )
-    }
-  })()
+	// Add genesis block and forked block to chain
+	const genesisBlockPromise = (async () => {
+		if (options.fork?.url) {
+			const block = await getBlockFromRpc(options.fork, options.common)
+			await putBlock(chain)(block)
+		} else {
+			await putBlock(chain)(
+				options.genesisBlock ?? createGenesisBlock(options.genesisStateRoot ?? EMPTY_STATE_ROOT, options.common),
+			)
+		}
+	})()
 
-  return chain
+	return chain
 }
